--- client/copy
@@ -1,35 +1,0 @@
-const Home = () => {
-  const { data, error } = useQuery(GET_CURRENT_USER);
-  console.log("home", data, error);
-
-  let posts;
-
-  if (data) {
-    const { loading: forUser, posts } = useQuery(GET_USER_POSTS);
-
-    console.log("i am the user", posts);
-  } else {
-    const { loading, posts } = useQuery(GET_All_POSTS);
-
-    console.log("theallpost", posts);
-  }
-
-  const store =  (
-    <Grid columns={3}>
-      <Grid.Row>
-        {loading ? (
-          <h1>loading posts...</h1>
-        ) : (
-          userposts &&
-          userposts.map((userpo) => (
-            <Grid.Column key={userpo.id}>
-              <PostCard res={userpo} />
-            </Grid.Column>
-          ))
-        )}
-      </Grid.Row>
-    </Grid>
-  ) ;
-
-  return store;
-};
