+++ client/src/Components/UpdatePost.js
@@ -1,27 +1,27 @@
-import React, { useCallback, useState } from "react";
+import { useCallback, useState } from "react";
 import { useMutation } from "@apollo/react-hooks";
-import { CreateNewPost } from "./graphql/mutation";
+import { EDIT_POST_BY_ID } from "../graphql/mutation";
 import { Form, Button } from "semantic-ui-react";
 
-export const Upload = () => {
+const UpdatePost = (props) => {
+  const id = props.match.params.id;
+  console.log(id);
   const [title, setTitle] = useState("le mee bebe bebe bebe ");
   const [description, setDescription] = useState(" hehe eee eege egeege");
   const [price, setPrice] = useState("");
-  const [image, setImage] = useState();
   const [category, setCategory] = useState("clothes");
-  const [file, setFile] = useState();
-  const id = localStorage.getItem("id");
 
-  const [createNewPost, { error, loading, data }] = useMutation(CreateNewPost, {
+  const [updatePost, { data, error }] = useMutation(EDIT_POST_BY_ID, {
     onError(error) {
       console.log("iiiiii", error);
     },
   });
+  console.log(error);
 
   return (
     <div className="regularInputs">
       <Form>
-        <h1>Create Post</h1>
+        <h1>Edit Post</h1>
         <input
           type="text"
           placeholder="title"
@@ -43,13 +43,7 @@ export const Upload = () => {
             setPrice(event.target.value);
           }}
         />
-        <input
-          type="file"
-          name="picture"
-          onChange={async (event) => {
-            await setFile(event.target.files[0]);
-          }}
-        />
+
         <select
           onChange={(event) => {
             setCategory(event.target.value);
@@ -64,26 +58,28 @@ export const Upload = () => {
 
         <div className="button">
           <button
-            onClick={() => {
-              const featureImage = [];
-              featureImage.push(file);
+            onClick={(e) => {
+              e.preventDefault();
+              console.log(id);
 
-              createNewPost({
+              updatePost({
                 variables: {
+                  id,
                   title,
                   description,
                   price,
                   category,
-                  featureImage,
-                  id,
                 },
               });
+              console.log(data);
             }}
           >
-            Create Post
+            Update Post
           </button>
         </div>
       </Form>
     </div>
   );
 };
+
+export default UpdatePost;
