--- client/src/context/auth.js
@@ -1,49 +1,0 @@
-import { useReducer, createContext } from "react";
-
-const AuthContext = createContext({
-  user: null,
-  login: (userData) => {},
-  logout: () => {},
-});
-
-function authReducer(state, action) {
-  switch (action.type) {
-    case "LOGIN":
-      return {
-        ...state,
-        user: action.payload,
-      };
-    case "LOGOUT":
-      return {
-        ...state,
-        user: null,
-      };
-    default:
-      return state;
-  }
-}
-
-function AuthProvider(props) {
-  const [state, dispatch] = useReducer(authReducer, { user: null });
-
-  function login(userData) {
-    dispatch({
-      type: "LOGIN",
-      payload: userData,
-    });
-  }
-
-  function logout(userData) {
-    dispatch({
-      type: "LOGOUT",
-    });
-  }
-  return (
-    <AuthContext.Provider
-      value={{ state: state.user, login, logout }}
-      {...props}
-    />
-  );
-}
-
-export { AuthContext, AuthProvider };
